backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pydantic-settings==2.0.3

# Test dependencies
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0
httpx==0.25.2
pytest-cov==4.1.0
factory-boy==3.3.0